+++ RayEngine/Source/DX11/DX11PipelineState.cpp
@@ -218,16 +218,16 @@ namespace RayEngine
 		void DX11PipelineState::SetName(const std::string& name)
 		{
 			std::string tmpName = name + " : InputLayout";
-			m_InputLayout->SetPrivateData(WKPDID_D3DDebugObjectName, tmpName.size(), tmpName.c_str());
+			m_InputLayout->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(tmpName.size()), tmpName.c_str());
 
 			tmpName = name + " : BlendState";
-			m_BlendState->SetPrivateData(WKPDID_D3DDebugObjectName, tmpName.size(), tmpName.c_str());
+			m_BlendState->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(tmpName.size()), tmpName.c_str());
 
 			tmpName = name + " : RasterizerState";
-			m_RasterizerState->SetPrivateData(WKPDID_D3DDebugObjectName, tmpName.size(), tmpName.c_str());
+			m_RasterizerState->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(tmpName.size()), tmpName.c_str());
 
 			tmpName = name + " : DepthStencilState";
-			m_DepthStencilState->SetPrivateData(WKPDID_D3DDebugObjectName, tmpName.size(), tmpName.c_str());
+			m_DepthStencilState->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(tmpName.size()), tmpName.c_str());
 		}
