+++ RayEngine/Source/DX11/DX11ShaderConstantBlock.cpp
@@ -71,7 +71,7 @@ namespace RayEngine
 		/////////////////////////////////////////////////////////////
 		void DX11ShaderConstantBlock::SetName(const std::string& name)
 		{
-			m_ConstantBuffer->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str());
+			m_ConstantBuffer->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32>(name.size()), name.c_str());
 		}
